# docs/OpenAPIv3_swagger.yaml
components:
  schemas:
...
    domain.Chat:
      properties:
        created_at:
          type: string
        id:
          type: integer
        messages:
          items:
            $ref: '#/components/schemas/domain.Message'
          type: array
          uniqueItems: false
        metadata:
          description: Store ChatMetadata as JSON string
          type: string
        organization_id:
          type: integer
        tags:
          description: JSON array of tags as string
          type: string
        title:
          type: string
        updated_at:
          type: string
        user_id:
          description: Nullable for anonymous chats
          type: integer
      required:
      - created_at
      - id
      - messages
      - metadata
      - organization_id
      - tags
      - title
      - updated_at
      - user_id
      type: object
    domain.ChatMetadata:
      description: Use the structured type
      properties:
        avg_response_time:
          description: In seconds
          type: number
        country_code:
          description: ISO-3166 country code (e.g., "NL")
          type: string
        ip_address:
          type: string
        is_escalated:
          type: boolean
        is_forwarded_to_hr:
          type: boolean
        language_code:
          description: ISO-639 language code (e.g., "tr")
          type: string
        question_category:
          type: string
        sentiment:
          type: string
        session_id:
          type: string
        token_count:
          description: Total tokens for the chat session
          type: integer
        transcript_link:
          type: string
        user_rating:
          description: Optional user rating
          type: integer
      required:
      - avg_response_time
      - country_code
      - ip_address
      - is_escalated
      - is_forwarded_to_hr
      - language_code
      - question_category
      - sentiment
      - session_id
      - token_count
      - transcript_link
      - user_rating
      type: object
...
    domain.Message:
      properties:
        chat_id:
          type: integer
        content:
          type: string
        created_at:
          type: string
        id:
          type: integer
        metadata:
          description: Store MessageMetadata as JSON string
          type: string
        role:
          $ref: '#/components/schemas/domain.MessageRole'
      required:
      - chat_id
      - content
      - created_at
      - id
      - metadata
      - role
      type: object
...
    domain.MessageRole:
      type: string
      x-enum-varnames:
      - MessageRoleUser
      - MessageRoleAssistant
      - MessageRoleSystem
...
    handler.CreateChatRequest:
      properties:
        metadata:
          $ref: '#/components/schemas/domain.ChatMetadata'
        tags:
          items:
            type: string
          type: array
          uniqueItems: false
        title:
          type: string
        user_id:
          description: Optional, for anonymous chats
          type: integer
      required:
      - metadata
      - tags
      - title
      - user_id
      type: object
...
    handler.GetChatResponse:
      properties:
        created_at:
          type: string
        id:
          type: integer
        messages:
          items:
            $ref: '#/components/schemas/domain.Message'
          type: array
          uniqueItems: false
        metadata:
          $ref: '#/components/schemas/domain.ChatMetadata'
        organization_id:
          type: integer
        tags:
          description: Parsed tags
          items:
            type: string
          type: array
          uniqueItems: false
        title:
          type: string
        updated_at:
          type: string
        user_id:
          description: Nullable for anonymous chats
          type: integer
      required:
      - created_at
      - id
      - messages
      - metadata
      - metadata
      - organization_id
      - tags
      - tags
      - title
      - updated_at
      - user_id
      type: object
...
    handler.UpdateChatRequest:
      properties:
        metadata:
          $ref: '#/components/schemas/domain.ChatMetadata'
        tags:
          items:
            type: string
          type: array
          uniqueItems: false
        title:
          type: string
      required:
      - metadata
      - tags
      - title
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: x-organization-api-key
      type: apiKey
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: chatlogger-api-go+swagger@kjanat.com
    name: ChatLogger
    url: https://github.com/kjanat/chatlogger-api-go/issues
  description: API for logging and managing chat sessions.
  license:
    name: MIT License
    url: https://github.com/kjanat/chatlogger-api-go/blob/master/LICENSE
  termsOfService: https://github.com/kjanat/chatlogger-api-go#terms-of-service
  title: ChatLogger API (Go)
  version: ""
openapi: 3.1.0
paths:
...
  /v1/chats:
    get:
      description: Retrieves a paginated list of chat sessions for the user's organization.
      parameters:
      - description: Number of chats per page
        in: query
        name: limit
        schema:
          default: 20
          type: integer
      - description: Offset for pagination
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/domain.Chat'
                type: array
          description: List of chats
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to list chats
      security:
      - BearerAuth: []
      summary: List Chats
      tags:
      - Chats
    post:
      description: Creates a new chat session for the organization. Can be called
        via Dashboard (JWT) or Public API (API Key).
      parameters:
      - description: Organization Slug (Required for Public API)
        in: path
        name: slug
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.CreateChatRequest'
        description: Chat Details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: 'message: Chat created successfully, chat_id: uint64'
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request data
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT or API Key invalid/missing, or Org ID not
            found)
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden (API Key doesn't match slug)
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to create chat or process tags/metadata
      security:
      - ApiKeyAuth: []
        BearerAuth: []
      summary: Create Chat
      tags:
      - Chats
  /v1/chats/{chatID}:
    delete:
      description: Deletes a chat session and its associated messages.
      parameters:
      - description: Chat ID
        in: path
        name: chatID
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Chat deleted successfully
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid chat ID
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Permission denied (Chat doesn't belong to user's org)
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Chat not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to get or delete chat
      security:
      - BearerAuth: []
      summary: Delete Chat
      tags:
      - Chats
    get:
      description: Retrieves details for a specific chat session, optionally including
        messages.
      parameters:
      - description: Chat ID
        in: path
        name: chatID
        required: true
        schema:
          type: integer
      - description: Include messages in the response
        in: query
        name: include_messages
        schema:
          default: false
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.GetChatResponse'
          description: Chat details
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid chat ID
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Permission denied (Chat doesn't belong to user's org)
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Chat not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to get chat or messages
      security:
      - BearerAuth: []
      summary: Get Chat by ID
      tags:
      - Chats
    patch:
      description: Updates the title, tags, or metadata of an existing chat session.
      parameters:
      - description: Chat ID
        in: path
        name: chatID
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.UpdateChatRequest'
        description: Fields to update
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.GetChatResponse'
          description: Updated chat details
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid chat ID or request data
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Permission denied (Chat doesn't belong to user's org)
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Chat not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to get or update chat, or process tags/metadata
      security:
      - BearerAuth: []
      summary: Update Chat
      tags:
      - Chats
...
  /v1/orgs/{slug}/chats:
    post:
      description: Creates a new chat session for the organization. Can be called
        via Dashboard (JWT) or Public API (API Key).
      parameters:
      - description: Organization Slug (Required for Public API)
        in: path
        name: slug
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.CreateChatRequest'
        description: Chat Details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: 'message: Chat created successfully, chat_id: uint64'
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request data
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT or API Key invalid/missing, or Org ID not
            found)
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden (API Key doesn't match slug)
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to create chat or process tags/metadata
      security:
      - ApiKeyAuth: []
        BearerAuth: []
      summary: Create Chat
      tags:
      - Chats
...